requirements.txt
asgiref==3.2.10
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
